Fixed jail issues
SetJail must check to add if they are already in a jail and cancel it.
Unjail must check all jails to see if anyone is in it.

To fix, implement a list in all jails which will be loaded onEnable and
on jail. Then check if the list contains the player.